Sentiment Analysis and Topic Modeling
SentimentAnalysis.py
Handles analysis for any comments we may not have an explicit rating for (e.g., metacritic). We could decide to just hand interpret these if the rating is over half, and this would be obsolete which is fine.

TopicModel.py
Creates a topic model using a minibatch kullback-leiber loss model as I found this was best suited for our limited dataset and handling potentially sparse input data. With this model, there is some final processing that can intake a comment and rating as a tuple. The comment is then mapped to which category it fits best (could expand to top 2 or 3 depending on what we see with real data). Then, depending on whether the comment was positive or negative, the topic's "rating" is summed. E.g. if a comment in topic one is positive, then it will return "1" as topic one's rating. However, if there were two negative comments and two positive comments most associated with topic 1 it will return "0" as the score as this is split information therefore more likely to differ player to player. Third case is if there are a lot of negative OR positive comments for a topic that get summed, then a topic will return with a large value in that section.

TLDR:
Creates 10 topics from training data and intakes a new comments (with pos or neg rating). Sorts out which topic each comment applies to and returns a summed "opinion" for each topic.